Remove unused conversion
