+++ .gitignore
@@ -1,2 +1,3 @@
 /target
 Cargo.lock
+/atspi/benches/resources/*
